Reconstitute temperature fields

Reconstitute temperature residual fields from the basis vectors and a matrix of coefficients. If meanfield is supplied, add the residual field to the mean field to get a reconstituted temperature field.

reconst_fields(basis, bcoord, meanfield = NULL)

Arguments

basis
A matrix [ngrid x N] of basis vectors, one vector per column.

bcoord
A matrix [ntime x N] of basis coordinates, one time step in each row, with coordinates across the columns.

meanfield
An optional matrix [ntime x ngrid] of mean field values, one time slice per row. If supplied, it will be added to the residuals calculated from basis and bcoord.

Value

A matrix [ntime x ngrid] of residual values (or temperature values, if meanfield was supplied).

Details

For a mean field generated using the default linear pattern scaling, one can get the mean field by running pscl_apply on the pscl member of a fldgen object.